Revert Changes for SR-6812
32-bit broke when Swift 4 introduced a change to how refcount bits were stored in swift objects. PR-1525 fixed this issue. SR-6812 was opened around the same time which is a dupe of the original issue, and PR-1557 attempted to fix this dupe in a different way.

The interaction of these two PRs re-broke 32-bit by shrinking the CFRuntimeBase to 12 bytes and breaking the 8-byte alignment it has historically had. This means it was unclear if the swift properties would start immediately after the CFRuntimeBase, or after some padding, meaning offsets into the object calculated manually are unreliable.

This change reverts PR-1557 to restore the 8-byte alignment of CFRuntimeBase on 32-bit platforms when using a swift deployment target.